La till en liten reader av propertyfilen i Roundklassen.
Organiserade strukturen av projektet till olika paket.
