+++ src/Quizkampen/Game/Game.java
@@ -1,4 +1,4 @@
-package Quizkampen;
+package Quizkampen.Game;
 
 public class Game {
