+++ src/Quizkampen/Model/Questions.java
@@ -1,4 +1,4 @@
-package Quizkampen;
+package Quizkampen.Model;
 
 public class Questions {
