+++ src/Quizkampen/Server/ClientHandler.java
@@ -1,4 +1,4 @@
-package Quizkampen;
+package Quizkampen.Server;
 
 import java.io.BufferedReader;
 import java.io.IOException;
