+++ src/game-config.properties
@@ -1,4 +1,4 @@
 # Game configurations for questions and rounds
 // TODO fixa så att det går att läsa av från denna fil och ställa in inställningarna beroende på värdena nedanför.
-questions.per.round = 2
-rounds.per.game = 2
+questions.per.round = 3
+rounds.per.game = 3
